requirements.txt
streamlit
feedparser
google-genai
textblob
cloudscraper
